feat(CardPreview): story 파일명 수정 및 카드사별 카드 색상 컴포넌트 추가
